requirements.txt
Django>=5.2.1
djangorestframework>=3.15
openai>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
